Protect (encrypted) root password: Do not provide it as argument
The encrypted root password was passed as argument to usermod. This
may enable other users to snoop the password by observing the running
processes. Especially if FAI softupdate runs are scheduled
periodically (e.g. on a daily basis). Even though the password appears
in its encrypted form, it should never appear in the list of running
processes.

Instead, use chpasswd with bash's "here strings" to protect the
encrypted root password. The chpasswd binary is part of shadow and
hence should be available everywhere.